client/requirements_dev.txt
pyserial==3.4
click==8.1.8
